requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
faiss-cpu
google-generativeai
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
langchain-classic
pypdf
pandas
numpy
unstructured
beautifulsoup4
